+++ tests/conftest.py
@@ -8,11 +8,12 @@ from .fixtures import hypothesis_strategies  # noqa: F401
 from .fixtures.auth import auth_credentials  # noqa: F401
 from .fixtures.backends import (  # noqa: F401
     es,
-    es2,
     es_data_stream,
+    es_forwarding,
     events,
+    lrs,
     mongo,
-    mongo2,
+    mongo_forwarding,
     swift,
     ws,
 )